feat: refactor runtime plugin
