+++ packages/runtime/plugin-runtime/src/core/appConfig.ts
@@ -5,7 +5,6 @@ const APP_CONFIG_SYMBOL = 'config';
 export const getConfig = (
   Component: React.ComponentType<any>,
 ): AppConfig | undefined =>
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-expect-error
   Component[APP_CONFIG_SYMBOL];
 
@@ -13,7 +12,6 @@ export const defineConfig = (
   Component: React.ComponentType<any>,
   config: AppConfig,
 ): React.ComponentType<any> => {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-expect-error
   Component[APP_CONFIG_SYMBOL] = config;
   return Component;
