+++ packages/runtime/plugin-runtime/src/core/loader/useLoader.ts
@@ -7,7 +7,7 @@ import {
   useEffect,
 } from 'react';
 import invariant from 'invariant';
-import { RuntimeReactContext } from '../../runtimeContext';
+import { RuntimeReactContext } from '../context/runtime';
 import { Loader, LoaderStatus } from './loaderManager';
 
 type LoaderFn<P = any, T = any> = (context: any, params: P) => Promise<T>;
