+++ packages/runtime/plugin-runtime/src/core/plugin/index.ts
@@ -1,0 +1,1 @@
+export * from './base';
