+++ packages/runtime/plugin-runtime/src/router/runtime/plugin.node.tsx
@@ -122,7 +122,7 @@ export const routerPlugin = ({
           if (routerContext instanceof Response) {
             // React Router would return a Response when redirects occur in loader.
             // Throw the Response to bail out and let the server handle it with an HTTP redirect
-            return routerContext;
+            return routerContext as any;
           }
 
           const router = createStaticRouter(routes, routerContext);
@@ -135,11 +135,11 @@ export const routerPlugin = ({
 
           return next({ context });
         },
-        hoc: ({ App }, next) => {
+        hoc: ({ App, config }, next) => {
           // can not get routes config, skip wrapping React Router.
           // e.g. App.tsx as the entrypoint
           if (!routesConfig) {
-            return next({ App });
+            return next({ App, config });
           }
 
           const getRouteApp = () => {
@@ -167,35 +167,13 @@ export const routerPlugin = ({
           if (routesConfig?.globalApp) {
             return next({
               App: hoistNonReactStatics(RouteApp, routesConfig.globalApp),
+              config,
             });
           }
 
           return next({
             App: RouteApp,
-          });
-        },
-        pickContext: ({ context, pickedContext }, next) => {
-          const { remixRouter } = context;
-
-          // remixRouter is not existed in conventional routes
-          if (!remixRouter) {
-            return next({ context, pickedContext });
-          }
-
-          // only export partial common API from remix-router
-          const router = {
-            navigate: remixRouter.navigate,
-            get location() {
-              return remixRouter.state.location;
-            },
-          };
-
-          return next({
-            context,
-            pickedContext: {
-              ...pickedContext,
-              router,
-            },
+            config,
           });
         },
       };
