+++ packages/runtime/plugin-runtime/src/router/runtime/plugin.tsx
@@ -74,11 +74,11 @@ export const routerPlugin = ({
 
           return next({ context });
         },
-        hoc: ({ App }, next) => {
+        hoc: ({ App, config }, next) => {
           // can not get routes config, skip wrapping React Router.
           // e.g. App.tsx as the entrypoint
           if (!finalRouteConfig && !createRoutes) {
-            return next({ App });
+            return next({ App, config });
           }
 
           const getRouteApp = () => {
@@ -180,36 +180,13 @@ export const routerPlugin = ({
           if (routesConfig?.globalApp) {
             return next({
               App: hoistNonReactStatics(RouteApp, routesConfig.globalApp),
+              config,
             });
           }
 
           return next({
             App: RouteApp,
-          });
-        },
-        pickContext: ({ context, pickedContext }, next) => {
-          const { remixRouter } = context;
-
-          // two scenarios: 1. remixRouter is not existed in conventional routes;
-          // 2. useRuntimeContext can be called by users before hoc hooks execute
-          if (!remixRouter) {
-            return next({ context, pickedContext });
-          }
-
-          // only export partial common API from remix-router
-          const router = {
-            navigate: remixRouter.navigate,
-            get location() {
-              return remixRouter.state.location;
-            },
-          };
-
-          return next({
-            context,
-            pickedContext: {
-              ...pickedContext,
-              router,
-            },
+            config,
           });
         },
       };
