+++ packages/runtime/plugin-runtime/src/ssr/index.tsx
@@ -136,31 +136,6 @@ export const ssr = (config: SSRPluginConfig): Plugin => ({
         context.ssrContext!.request = formatClient(request);
         return next({ context });
       },
-      pickContext: ({ context, pickedContext }, next) => {
-        const request: SSRServerContext['request'] | undefined =
-          window?._SSR_DATA?.context?.request;
-        const { initialData } = context;
-
-        if (!request) {
-          return next({
-            context,
-            pickedContext: {
-              ...pickedContext,
-              initialData,
-            },
-          });
-        }
-
-        return next({
-          context,
-          pickedContext: {
-            ...pickedContext,
-            initialData,
-            request,
-            response: mockResp,
-          },
-        });
-      },
     };
   },
 });
